+++ CHANGES.md
@@ -1,4 +1,4 @@
-# Changelog 
+# Changelog
 
 
 ## in progress
@@ -7,3 +7,4 @@
 ## 2023-08-xx 0.0.0
 - Initial thing
 - Journey planning basics and HTML output
+- Fix return journey wrt. arrival time